util: Remove unnecessary g from is_valid_uuid regex.
prep-commit

1- No need to do a global match
(i.e. find all matches). Also in the case of /g,
the regex object keeps track of the last match position,
this means we have to reset it before each call
to is_valid_uuid.test()

2- Move "is_valid_uuid" from "remote_billing_auth"
to "util.ts" because it will be used again
by other files in the next commit.